requirements.txt
langchain==0.3.14
langchain-google-genai==2.0.7
langchain-core==0.3.29
google-generativeai==0.8.3
python-dotenv==1.0.1
